requirements.txt
# 宝可梦Like游戏多人机制设计演进研究工具
# Pokemon-like Game Multiplayer Design Evolution Research Tool

streamlit>=1.28.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
langchain>=0.1.0
langchain-openai>=0.0.5
langchain-anthropic>=0.1.0
chromadb>=0.4.0
plotly>=5.18.0
pandas>=2.1.0
python-dotenv>=1.0.0

# 可选依赖（用于增强社区反应数据采集）
# praw>=7.0.0   # Reddit API 客户端，配置后采集效果更稳定
steamreviews>=0.9.6   # Steam 评测免费采集，无需 API Key
